+++ src/main/java/durikkiri/project/entity/dto/auth/RefreshTokenInfoDto.java
@@ -1,0 +1,14 @@
+package durikkiri.project.entity.dto.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class RefreshTokenInfoDto {
+
+    private String account;
+    private String refreshToken;
+    private String authorities;
+
+}
